+++ src/config.py
@@ -1,15 +1,13 @@
 import os
 
-S3_LOCATION = 'http://your-amazon-site.amazonaws.com/'
-S3_KEY = 'YOURAMAZONKEY'
-S3_SECRET = 'YOURAMAZONSECRET'
-S3_UPLOAD_DIRECTORY = 'what_directory_on_s3'
-S3_BUCKET = 'transcriptsdevx'
+BOTO3_ACCESS_KEY = os.environ['AWS_ACCESS_KEY_ID']
+BOTO3_SECRET_KEY= os.environ['AWS_SECRET_ACCESS_KEY']
+BOTO3_REGION= 'us-west-2'
+BOTO3_SERVICES =['dynamodb']
 
 UPLOAD_FOLDER = 'uploads'
 ALLOWED_EXTENSIONS = set(['pdf'])
 MAX_CONTENT_LENGTH = 16 * 1024 * 1024 #max file size is 16MB
 
-SECRET_KEY = "FLASK_SECRET_KEY"
 DEBUG = True
 PORT = 5000
